refactor: organize Python sandbox packages
