+++ data-agent-management/src/main/java/com/alibaba/cloud/ai/dataagent/service/code/sandbox/SaaSandboxPythonCodeExecutorService.java
@@ -17,6 +17,7 @@ package com.alibaba.cloud.ai.dataagent.service.code.sandbox;
 
 import com.alibaba.cloud.ai.dataagent.properties.CodeExecutorProperties;
 import com.alibaba.cloud.ai.dataagent.service.code.PythonCodeExecutorService;
+import com.alibaba.cloud.ai.dataagent.service.code.sandbox.runtime.SaaSandboxTaskRunner;
 import jakarta.annotation.PreDestroy;
 import org.springframework.stereotype.Service;
