+++ data-agent-management/src/main/java/com/alibaba/cloud/ai/dataagent/service/code/sandbox/runtime/SaaSandboxRuntime.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.cloud.ai.dataagent.service.code.sandbox;
+package com.alibaba.cloud.ai.dataagent.service.code.sandbox.runtime;
 
 import com.alibaba.cloud.ai.dataagent.properties.CodeExecutorProperties;
 import io.agentscope.runtime.sandbox.box.BaseSandbox;
